create component container and component slide
