+++ demo/index.html
@@ -5,11 +5,11 @@
   <title>slider-carousel demo</title>
   <meta charset="utf-8" />
   <meta name="viewport" content="width=device-width, minimum-scale=1.0, initial-scale=1, user-scalable=yes" />
-  <script type="module" src="../components/parent-element.js"></script>
+  <script type="module" src="../components/carousel-container.js"></script>
 </head>
 
 <body>
-  <parent-element></parent-element>
+  <carousel-container></carousel-container>
 </body>
 
 </html>
